requirements.txt
apispec==1.3.3
attrs==19.3.0
Babel==2.8.0
bcrypt==3.1.7
blinker==1.4
bottle==0.12.18
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
defusedxml==0.6.0
dnspython==1.16.0
Flask==1.1.1
Flask-AppBuilder==2.2.4
Flask-Babel==1.0.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-DebugToolbar==0.11.0
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-MongoAlchemy==0.7.2
flask-mongoengine==0.9.5
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
ibm-cloud-sdk-core==1.5.1
ibm-watson==4.3.0
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.22.2
MongoAlchemy==0.19
mongoengine==0.19.1
prison==0.1.2
pycparser==2.19
PyJWT==1.7.1
# pymongo==3.9.0
pymongo[tls,srv]==3.6.1
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.11.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.13
SQLAlchemy-Utils==0.36.1
urllib3